fix HH delivery when column index is larger than 2GB patch by jbellis; reviewed by slebresne for CASSANDRA-2297
git-svn-id: https://svn.apache.org/repos/asf/cassandra/branches/cassandra-0.7@1080716 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/cassandra/io/util/FileDataInput.java
@@ -38,7 +38,7 @@ public interface FileDataInput extends DataInput, Closeable
 
     public void reset(FileMark mark) throws IOException;
 
-    public int bytesPastMark(FileMark mark);
+    public long bytesPastMark(FileMark mark);
 
     /**
      * Read length bytes from current file position
